backend/pyproject.toml
[tool.poetry]
name = "storage-unit"
version = "0.1.0"
description = ""
authors = ["MuhammadHassan1998"]
readme = "README.md"
packages = [{include = "storage_unit"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
black = "23.3.0"
flake8 = "6.0.0"
python-dotenv = "1.0.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
six = "1.16.0"
django-cors-headers = "3.14.0"
xlrd = "2.0.1"
pandas = "2.0.1"
openpyxl = "3.1.2"
asgiref = "3.6.0"
click = "8.1.3"
django-leaflet = "0.29.0"
et-xmlfile = "1.1.0"
geographiclib = "2.0"
mccabe = "0.7.0"
mypy-extensions = "1.0.0"
numpy = "1.24.3"
packaging = "23.1"
pathspec = "0.11.1"
platformdirs = "3.5.0"
psycopg2-binary = "2.9.6"
pycodestyle = "2.10.0"
pyflakes = "3.0.1"
pyjwt = "2.6.0"
python-dateutil = "2.8.2"
pytz = "2023.3"
sqlparse = "0.4.4"
tomli = "2.0.1"
tzdata = "2023.3"
certifi = "2023.5.7"
charset-normalizer = "3.1.0"
django-filter = "23.2"
idna = "3.4"
requests = "2.31.0"
urllib3 = "2.0.2"
pillow = "9.5.0"
amqp = "5.1.1"
async-timeout = "4.0.3"
billiard = "4.1.0"
celery = "5.3.1"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
gunicorn = "21.2.0"
kombu = "5.3.1"
prompt-toolkit = "3.0.39"
redis = "5.0.0"
vine = "5.0.0"
wcwidth = "0.2.6"
attrs = "23.1.0"
autobahn = "23.1.2"
automat = "22.10.0"
cffi = "1.15.1"
channels = "4.0.0"
channels-redis = "4.1.0"
constantly = "15.1.0"
cryptography = "41.0.3"
daphne = "4.0.0"
django-jazzmin = "2.6.0"
hyperlink = "21.0.0"
incremental = "22.10.0"
msgpack = "1.0.5"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyopenssl = "23.2.0"
service-identity = "23.1.0"
twisted = "23.8.0"
txaio = "23.1.1"
typing-extensions = "4.7.1"
zope-interface = "6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
